general: move string query from api layer to dao layer
