+++ dao/common.go
@@ -4,10 +4,12 @@ import (
 	"context"
 	"database/sql"
 	"errors"
+	"fmt"
 	"strings"
 
 	"github.com/geerew/off-course/database"
 	"github.com/geerew/off-course/utils"
+	"github.com/geerew/off-course/utils/queryparser"
 	"github.com/geerew/off-course/utils/types"
 )
 
@@ -187,4 +189,34 @@ func sanitizeIDs(in []string) []string {
 		out = append(out, id)
 	}
 	return out
+}
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+// ErrStringQueryParse is returned when the `q` parameter fails to parse
+var ErrStringQueryParse = errors.New("list query parse error")
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+// applyStringQuerySortOnly parses dbOpts.StringQuery when Query is non-empty and applies sort tokens only
+// (no WHERE from the expression). Used for lesson and attachment lists that support `q` ordering.
+func applyStringQuerySortOnly(dbOpts *Options) error {
+	if dbOpts == nil || dbOpts.StringQuery == nil || dbOpts.StringQuery.Query == "" {
+		return nil
+	}
+
+	parsed, err := queryparser.Parse(dbOpts.StringQuery.Query, dbOpts.StringQuery.AllowedFilters)
+	if err != nil {
+		return fmt.Errorf("%w: %w", ErrStringQueryParse, err)
+	}
+
+	if parsed == nil {
+		return nil
+	}
+
+	if len(parsed.Sort) > 0 {
+		dbOpts.OverrideOrderBy(parsed.Sort...)
+	}
+
+	return nil
 }
