+++ dao/user.go
@@ -2,10 +2,12 @@ package dao
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/Masterminds/squirrel"
 	"github.com/geerew/off-course/models"
 	"github.com/geerew/off-course/utils"
+	"github.com/geerew/off-course/utils/queryparser"
 )
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -73,6 +75,10 @@ func (dao *DAO) GetUser(ctx context.Context, dbOpts *Options) (*models.User, err
 // ListUsers gets all records from the user table based upon the where clause and pagination
 // in the options
 func (dao *DAO) ListUsers(ctx context.Context, dbOpts *Options) ([]*models.User, error) {
+	if err := applyUsersStringQuery(dbOpts); err != nil {
+		return nil, err
+	}
+
 	builderOpts := newBuilderOptions(models.USER_TABLE).
 		WithColumns(models.UserColumns()...).
 		SetDbOpts(dbOpts)
@@ -130,4 +136,67 @@ func (dao *DAO) DeleteUsers(ctx context.Context, dbOpts *Options) error {
 
 	_, err := dao.db.ExecContext(ctx, sqlStr, args...)
 	return err
+}
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+// applyUsersStringQuery parses dbOpts.StringQuery when Query is non-empty and sets Where / OrderBy.
+func applyUsersStringQuery(dbOpts *Options) error {
+	if dbOpts == nil || dbOpts.StringQuery == nil || dbOpts.StringQuery.Query == "" {
+		return nil
+	}
+
+	parsed, err := queryparser.Parse(dbOpts.StringQuery.Query, dbOpts.StringQuery.AllowedFilters)
+	if err != nil {
+		return fmt.Errorf("%w: %w", ErrStringQueryParse, err)
+	}
+
+	if parsed == nil {
+		return nil
+	}
+
+	if len(parsed.Sort) > 0 {
+		dbOpts.OverrideOrderBy(parsed.Sort...)
+	}
+
+	dbOpts.WithWhere(usersWhereBuilder(parsed.Expr))
+
+	return nil
+}
+
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+// usersWhereBuilder builds a squirrel WHERE expression from a queryparser.QueryExpr
+func usersWhereBuilder(expr queryparser.QueryExpr) squirrel.Sqlizer {
+	switch node := expr.(type) {
+	case *queryparser.ValueExpr:
+		return squirrel.Or{
+			squirrel.Like{"LOWER(" + models.USER_TABLE_USERNAME + ")": "%" + node.Value + "%"},
+			squirrel.Like{"LOWER(" + models.USER_TABLE_DISPLAY_NAME + ")": "%" + node.Value + "%"},
+		}
+	case *queryparser.FilterExpr:
+		switch node.Key {
+		case "role":
+			return squirrel.Eq{models.USER_TABLE_ROLE: node.Value}
+
+		default:
+			return nil
+		}
+	case *queryparser.AndExpr:
+		var andSlice []squirrel.Sqlizer
+		for _, child := range node.Children {
+			andSlice = append(andSlice, usersWhereBuilder(child))
+		}
+
+		return squirrel.And(andSlice)
+	case *queryparser.OrExpr:
+		var orSlice []squirrel.Sqlizer
+		for _, child := range node.Children {
+			orSlice = append(orSlice, usersWhereBuilder(child))
+		}
+
+		return squirrel.Or(orSlice)
+	default:
+		return nil
+	}
 }
